[DEV-2992] moving IDV tooltip file to shared dir
